+++ cli/src/commands/gerrit/upload.rs
@@ -24,6 +24,7 @@ use jj_lib::commit::Commit;
 use jj_lib::git;
 use jj_lib::git::GitPushOptions;
 use jj_lib::git::GitRefUpdate;
+use jj_lib::git::GitSubprocessCallback;
 use jj_lib::git::GitSubprocessOptions;
 use jj_lib::merge::Diff;
 use jj_lib::object_id::ObjectId as _;
@@ -33,6 +34,7 @@ use jj_lib::settings::UserSettings;
 use jj_lib::store::Store;
 use jj_lib::trailer::Trailer;
 use jj_lib::trailer::parse_description_trailers;
+use regex::Regex;
 
 use crate::cli_util::CommandHelper;
 use crate::cli_util::RevisionArg;
@@ -47,17 +49,13 @@ use crate::ui::Ui;
 
 /// Upload changes to Gerrit for code review, or update existing changes.
 ///
-/// Uploading in a set of revisions to Gerrit creates a single "change" for
+/// Uploading a set of revisions to Gerrit creates a single "change" for
 /// each revision included in the revset. These changes are then available
 /// for review on your Gerrit instance.
 ///
 /// Note: The Gerrit commit Id may not match that of your local commit Id,
-/// since we add a `Change-Id` footer to the commit message if one does not
-/// already exist. This ID is based off the jj Change-Id, but is not the same.
-///
-/// If a change already exists for a given revision (i.e. it contains the
-/// same `Change-Id`), this command will update the contents of the existing
-/// change to match.
+/// since we add a footer to the commit message if one does not
+/// already exist.
 ///
 /// Note: this command takes 1-or-more revsets arguments, each of which can
 /// resolve to multiple revisions; so you may post trees or ranges of
@@ -373,6 +371,70 @@ fn push_options(args: &UploadArgs) -> Result<Vec<String>, CommandError> {
     .collect())
 }
 
+const GERRIT_HOST_REGEX: &str = r"^  (https?://[^ ]*)/c/[^+]+\+/\d+\b";
+
+fn calculate_gerrit_host(lines: &[Vec<u8>]) -> Option<String> {
+    // See https://gerrit-review.googlesource.com/Documentation/intro-gerrit-walkthrough.html#_creating_the_review
+    // It's unclear looking from the docs if the output format is standardized,
+    // so we just take a best-effort approach.
+    // Example output:
+    // remote: Processing changes: new: 1, done
+    // remote:
+    // remote: New Changes:
+    // remote:   http://gerrithost/#/c/RecipeBook/+/702 Change to a proper, yeast based pizza dough.
+    // remote:
+    let gerrit_host = Regex::new(GERRIT_HOST_REGEX).unwrap();
+    lines
+        .iter()
+        .filter_map(|line| std::str::from_utf8(line).ok())
+        .filter_map(|line| gerrit_host.captures(line))
+        .map(|captures| captures[1].to_string())
+        .next()
+}
+
+// TeeCallback acts like the "tee" command in unix.
+// It prints to the UI, but also captures the remote sideband output from git.
+struct TeeCallback<'a> {
+    ui: GitSubprocessUi<'a>,
+    lines: Vec<Vec<u8>>,
+}
+
+impl<'a> TeeCallback<'a> {
+    fn new(ui: &'a Ui) -> Self {
+        Self {
+            ui: GitSubprocessUi::new(ui),
+            lines: vec![],
+        }
+    }
+}
+
+impl GitSubprocessCallback for TeeCallback<'_> {
+    fn needs_progress(&self) -> bool {
+        self.ui.needs_progress()
+    }
+
+    fn progress(&mut self, progress: &git::GitProgress) -> std::io::Result<()> {
+        self.ui.progress(progress)
+    }
+
+    fn local_sideband(
+        &mut self,
+        message: &[u8],
+        term: Option<git::GitSidebandLineTerminator>,
+    ) -> std::io::Result<()> {
+        self.ui.local_sideband(message, term)
+    }
+
+    fn remote_sideband(
+        &mut self,
+        message: &[u8],
+        term: Option<git::GitSidebandLineTerminator>,
+    ) -> std::io::Result<()> {
+        self.lines.push(message.to_vec());
+        self.ui.remote_sideband(message, term)
+    }
+}
+
 pub async fn cmd_gerrit_upload(
     ui: &mut Ui,
     command: &CommandHelper,
@@ -496,6 +558,7 @@ pub async fn cmd_gerrit_upload(
         }
     }
 
+    let mut review_url = command.settings().get_string("gerrit.review-url").ok();
     let mut old_to_new: HashMap<CommitId, Commit> = HashMap::new();
     for original_commit in to_upload.into_iter().rev() {
         let trailers = parse_description_trailers(original_commit.description());
@@ -528,9 +591,12 @@ pub async fn cmd_gerrit_upload(
                     short_change_hash(original_commit.change_id()),
                 )?;
             }
-            if trailer.key == "Link"
-                && !matches!(trailer.value.split_once("/id/I"), Some((_url, id)) if id.len() == 40)
-            {
+            let is_link_valid = if let Some((_url, id)) = trailer.value.split_once("/id/I") {
+                id.len() == 40
+            } else {
+                false
+            };
+            if trailer.key == "Link" && !is_link_valid {
                 writeln!(
                     ui.warning_default(),
                     "Invalid Link footer in revision {}",
@@ -629,6 +695,8 @@ pub async fn cmd_gerrit_upload(
 
         let new_commit = old_to_new.get(head).unwrap();
 
+        let mut callback = TeeCallback::new(ui);
+
         // how do we get better errors from the remote? 'git push' tells us
         // about rejected refs AND ALSO '(nothing changed)' when there are no
         // changes to push, but we don't get that here.
@@ -640,7 +708,7 @@ pub async fn cmd_gerrit_upload(
                 qualified_name: remote_ref.clone().into(),
                 targets: Diff::new(None, Some(new_commit.id().clone())),
             }],
-            &mut GitSubprocessUi::new(ui),
+            &mut callback,
             &push_options,
         )
         // Despite the fact that a manual git push will error out with 'no new
@@ -660,7 +728,23 @@ pub async fn cmd_gerrit_upload(
         if !push_stats.all_ok() {
             return Err(user_error("Failed to push all changes to gerrit"));
         }
+
+        if review_url.is_none() {
+            // Ensure that on the next iteration of the loop, we don't re-output this.
+            review_url = calculate_gerrit_host(&callback.lines);
+            if let Some(review_url) = review_url.as_deref() {
+                writeln!(
+                    ui.warning_default(),
+                    "Gerrit URL is not set, some features will not work"
+                )?;
+                writeln!(
+                    ui.hint_default(),
+                    "Run `jj config set --repo gerrit.review-url {review_url}",
+                )?;
+            }
+        }
     }
+
     Ok(())
 }
 
@@ -760,6 +844,35 @@ mod tests {
             .into_iter()
             .map(|s| s.to_string())
             .collect::<Vec<_>>(),
+        );
+    }
+
+    #[test]
+    fn gerrit_host_regex_test() {
+        let re = Regex::new(GERRIT_HOST_REGEX).unwrap();
+        let parse = |x| re.captures(x).map(|c| c[1].to_string());
+        assert_eq!(
+            parse(
+                "  http://gerrithost/#/c/RecipeBook/+/702 Change to a proper, yeast based pizza \
+                 dough."
+            ),
+            Some("http://gerrithost/#".to_string())
+        );
+        assert_eq!(
+            parse(
+                "  http://gerrithost/#/RecipeBook/+/702 Change to a proper, yeast based pizza \
+                 dough."
+            ),
+            None
+        );
+        assert_eq!(
+            parse("  http://gerrithost/#/c/+/702 Change to a proper, yeast based pizza dough."),
+            None
+        );
+        // A real-world example.
+        assert_eq!(
+            parse("  https://chromium-review.googlesource.com/c/build/+/123456 hello [NEW]"),
+            Some("https://chromium-review.googlesource.com".to_string())
         );
     }
 }